Fix logical race with checking closed and is empty.
Previously the check for both vars was checked with a mutex, however
by removing the public mutex the values are unsynchronized.

There could be a case where concurrently:

1. queue is empty
2. process goroutine reads IsEmpty == true
3. item is added to queue
4. queue is closed
5. process goroutine reads IsClosed == true

If we reorder the IsClosed and IsEmpty then this won't be a problem.
However, there probably are clearer ways to avoid this problem.