core: simplify the process and network startup code
When the interfaces are properly aligned, we don't need to do any
casting.

Similarly, by using `isDrained` simplifies the logic for checking
whether the process should be still running. isDrained returns true when
the connection is closed and there isn't anything more to read.

Adding a method `isSelfStarting` it simplifies the startup code.