core: implement deadlock detection
Tracking per process status for deadlock detection is trivial, however
we need to account for connection where there are suspended receivers
and senders.

To properly handle this scenario we introduce a concept of "live
connections". Connection is considered live when there could be further
progress.

Connection becomes live when:
1. There are both pending receives and sends or
2. Connection is closed and there is a pending receiver.

Connection becomes unlive when:
1. There are no more receivers or
2. There are no more pending sends and the connection is empty.